# docs/swagger.yaml
definitions:
...
  models.Response:
    properties:
      body: {}
      message:
        type: string
      status:
        type: boolean
    type: object
...
  models.SupplierCreate:
    properties:
      address:
        type: string
      email:
        type: string
      name:
        type: string
      phone:
        type: string
    required:
    - name
    type: object
  models.SupplierLaundry:
    properties:
      address:
        type: string
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      phone:
        type: string
      updated_at:
        type: string
    type: object
  models.SupplierUpdate:
    properties:
      address:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      phone:
        type: string
    required:
    - id
    - name
    type: object
...
info:
  contact: {}
  description: This is a api to app gestioncar
  termsOfService: http://swagger.io/terms/
  title: APP GESTIONCAR
  version: "1.0"
paths:
...
  /supplier/{id}:
    get:
      consumes:
      - application/json
      description: Get a supplier by its ID within a specified workplace.
      parameters:
      - description: Workplace Token
        in: header
        name: X-Workplace-Token
        required: true
        type: string
      - description: ID of the supplier
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Supplier obtained with success
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                body:
                  $ref: '#/definitions/models.SupplierLaundry'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "401":
          description: Auth is required
          schema:
            $ref: '#/definitions/models.Response'
        "403":
          description: Not Authorized
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Supplier not found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - BearerAuth: []
      summary: Get Supplier By ID
      tags:
      - Supplier
  /supplier/create:
    post:
      consumes:
      - application/json
      description: Creates a new supplier within the specified workplace.
      parameters:
      - description: Workplace Token
        in: header
        name: X-Workplace-Token
        required: true
        type: string
      - description: Details of the supplier to create
        in: body
        name: supplier
        required: true
        schema:
          $ref: '#/definitions/models.SupplierCreate'
      produces:
      - application/json
      responses:
        "200":
          description: Supplier created successfully
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                body:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "401":
          description: Auth is required
          schema:
            $ref: '#/definitions/models.Response'
        "403":
          description: Not Authorized
          schema:
            $ref: '#/definitions/models.Response'
        "422":
          description: Model is invalid
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - BearerAuth: []
      summary: Create Supplier
      tags:
      - Supplier
  /supplier/delete/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes a supplier based on the provided ID and workplace context.
      parameters:
      - description: Workplace Token
        in: header
        name: X-Workplace-Token
        required: true
        type: string
      - description: ID of the supplier
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Supplier deleted with success
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "401":
          description: Auth is required
          schema:
            $ref: '#/definitions/models.Response'
        "403":
          description: Not Authorized
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Supplier not found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - BearerAuth: []
      summary: Delete Supplier
      tags:
      - Supplier
  /supplier/get_all:
    get:
      consumes:
      - application/json
      description: Get All Suppliers
      parameters:
      - description: Workplace Token
        in: header
        name: X-Workplace-Token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Suppliers obtained with success
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                body:
                  items:
                    $ref: '#/definitions/models.SupplierLaundry'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "401":
          description: Auth is required
          schema:
            $ref: '#/definitions/models.Response'
        "403":
          description: Not Authorized
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - BearerAuth: []
      summary: Get All Suppliers
      tags:
      - Supplier
  /supplier/get_by_name:
    get:
      consumes:
      - application/json
      description: Fetches suppliers from either laundry or workshop based on the
        provided name and workplace.
      parameters:
      - description: Workplace Token
        in: header
        name: X-Workplace-Token
        required: true
        type: string
      - description: Name of the Supplier
        in: query
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of suppliers
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                body:
                  items:
                    $ref: '#/definitions/models.SupplierLaundry'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "401":
          description: Auth is required
          schema:
            $ref: '#/definitions/models.Response'
        "403":
          description: Not Authorized
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - BearerAuth: []
      - BearerAuth: []
      summary: Get Supplier By Name
      tags:
      - Supplier
  /supplier/update/{id}:
    put:
      consumes:
      - application/json
      description: Update a supplier's information from the specified workplace.
      parameters:
      - description: Workplace Token
        in: header
        name: X-Workplace-Token
        required: true
        type: string
      - description: ID of the Supplier
        in: path
        name: id
        required: true
        type: string
      - description: Supplier information
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.SupplierUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: Supplier updated with success
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "401":
          description: Auth is required
          schema:
            $ref: '#/definitions/models.Response'
        "403":
          description: Not Authorized
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Supplier not found
          schema:
            $ref: '#/definitions/models.Response'
        "422":
          description: Model is invalid
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - BearerAuth: []
      summary: Update Supplier
      tags:
      - Supplier
...
  /workplace/get_all:
    get:
      consumes:
      - application/json
      description: Workplace GetAll required auth token
      produces:
      - application/json
      responses:
        "200":
          description: Workplaces obtenidos con éxito
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "401":
          description: Auth is required
          schema:
            $ref: '#/definitions/models.Response'
        "403":
          description: Not Authorized
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - BearerAuth: []
      summary: Workplace GetAll
      tags:
      - Workplace
securityDefinitions:
  BearerAuth:
    description: 'Type "Bearer" followed by a space and the JWT token. Example: "Bearer
      eyJhbGciOiJIUz..."'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"